# openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "example": {
          "detail": "top_k must be between 1 and 50.",
          "instance": "/api/v1/movies/search",
          "parameter": "top_k",
          "status": 400,
          "title": "Invalid request",
          "traceId": "00-e79e5b5e8f35cb27aa25fbdd351befff-991094dec4ca84c6-01",
          "type": "https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.1"
        },
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
      "TokenRequest": {
        "description": "A client-credentials token request.",
        "example": {
          "client_id": "reader-client",
          "client_secret": "reader-secret-change-me",
          "grant_type": "client_credentials"
        },
        "properties": {
          "client_id": {
            "description": "Client identifier.",
            "type": "string"
          },
          "client_secret": {
            "description": "Client secret.",
            "type": "string"
          },
          "grant_type": {
            "description": "OAuth 2.0 grant type. Only `client_credentials` is supported. The field\nmay be omitted, in which case `client_credentials` is assumed.",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "client_id",
          "client_secret"
        ],
        "type": "object"
      },
      "TokenResponse": {
        "description": "A successfully issued access token.",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJyZWFkZXIt...",
          "expires_in": 3600,
          "roles": [
            "reader"
          ],
          "token_type": "Bearer"
        },
        "properties": {
          "access_token": {
            "description": "The signed JWT. Send it as `Authorization: Bearer &lt;token&gt;`.",
            "type": "string"
          },
          "expires_in": {
            "description": "Lifetime of the token in seconds.",
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "roles": {
            "description": "Roles the token grants.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "token_type": {
            "description": "Always `Bearer`.",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "roles"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "A JWT obtained from `POST /auth/token`. Send it as `Authorization: Bearer <token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Movie Search Platform"
    },
    "description": "Semantic search over a movie catalogue.\n\nThe API translates a natural language query into a vector, narrows the catalogue with the metadata filters supplied, and returns the closest matches with their similarity scores. The search itself runs on a Python MCP server backed by PostgreSQL with pgvector.\n\n**Getting a token.** Every `/api/v1/*` route needs a bearer token. Post your client id and secret to `/auth/token`, then send the returned token as `Authorization: Bearer <token>`.\n\n**Roles.** A `reader` token may search and read movies. A `admin` token may do that and also read `/api/v1/stats`.\n\n**Rate limit.** 60 requests per minute per client. A rejected request answers 429 and carries a `Retry-After` header.",
    "license": {
      "name": "MIT"
    },
    "title": "Intelligent Movie Search API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/auth/token": {
      "post": {
        "description": "Implements the OAuth 2.0 client-credentials grant. Send the client id and secret in a JSON body. The returned token goes in the `Authorization` header as `Bearer <token>` on every `/api/v1/*` request.",
        "operationId": "IssueToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Exchange client credentials for an access token.",
        "tags": [
          "Authentication"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:58080/"
    }
  ],
...
}